backend/requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
pydantic==2.12.0

sqlmodel==0.0.18 
# psycopg==3.1.8 
bcrypt==4.0.1
# passlib[bcrypt]==1.7.4
passlib
python-jose[cryptography]==3.3.0
psycopg2-binary
python-multipart